made some fixes with switch theme logic, remove div-container and other tiny things
