+++ src/app.tsx
@@ -1,7 +1,7 @@
 import { store } from 'src/store';
 import { Provider as StoreProvider } from 'react-redux';
 import Router from 'src/features/router';
-import ThemeContextProvider from 'src/components/switchTheme/useSwitchTheme';
+import { ThemeContextProvider } from 'src/components/switchTheme/useThemeContext';
 
 import './app.css';
