+++ src/components/menu/portal.test.tsx
@@ -5,6 +5,7 @@ import { Portal } from './portal';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 import { MemoryRouter } from 'react-router-dom';
+import { ThemeContext } from 'src/components/switchTheme/useThemeContext';
 
 vi.mock('react-dom', async () => {
   const actual = await vi.importActual<typeof import('react-dom')>('react-dom');
@@ -29,6 +30,12 @@ const store = configureStore({
   },
 });
 
+const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
+  const themeContextValue = { theme: 'light' };
+
+  return <ThemeContext.Provider value={themeContextValue}>{children}</ThemeContext.Provider>;
+};
+
 describe('Portal Component', () => {
   let portalRoot: HTMLDivElement;
 
@@ -46,9 +53,11 @@ describe('Portal Component', () => {
     const onClose = vi.fn();
     render(
       <Provider store={store}>
-        <MemoryRouter>
-          <Portal onClose={onClose} />
-        </MemoryRouter>
+        <ThemeProvider>
+          <MemoryRouter>
+            <Portal onClose={onClose} />
+          </MemoryRouter>
+        </ThemeProvider>
       </Provider>,
     );
 
@@ -61,9 +70,11 @@ describe('Portal Component', () => {
     const onClose = vi.fn();
     render(
       <Provider store={store}>
-        <MemoryRouter>
-          <Portal onClose={onClose} />
-        </MemoryRouter>
+        <ThemeProvider>
+          <MemoryRouter>
+            <Portal onClose={onClose} />
+          </MemoryRouter>
+        </ThemeProvider>
       </Provider>,
     );
