+++ src/components/switchTheme/switchTheme.tsx
@@ -1,6 +1,6 @@
 import { IconicButton } from 'src/components/_common/iconicButton/iconicButton';
 import { SwitchThemeIcon } from 'src/components/_common/icons/switchThemeIcon';
-import { useTheme } from 'src/components/switchTheme/useSwitchTheme';
+import { useTheme } from 'src/components/switchTheme/hook';
 
 import { useSelector } from 'react-redux';
 import { selectMenu } from 'src/store/selectors/menuButton';
