+++ src/components/switchTheme/useThemeContext.tsx
@@ -1,6 +1,6 @@
-import { createContext, ReactNode, useCallback, useContext, useEffect, useState } from 'react';
+import { createContext, ReactNode, useEffect, useState } from 'react';
 
-interface ThemeContextType {
+export interface ThemeContextType {
   theme: 'light' | 'dark';
   toggleTheme: () => void;
 }
@@ -11,7 +11,7 @@ interface IProps {
 
 export const ThemeContext = createContext<ThemeContextType | null>(null);
 
-const ThemeContextProvider = ({ children }: IProps) => {
+export const ThemeContextProvider = ({ children }: IProps) => {
   const getInitialTheme = () => {
     const savedTheme = localStorage.getItem('theme');
     if (savedTheme) {
@@ -24,14 +24,14 @@ const ThemeContextProvider = ({ children }: IProps) => {
 
   const [theme, setTheme] = useState<'light' | 'dark'>(getInitialTheme);
 
-  const toggleTheme = useCallback(() => {
+  const toggleTheme = () => {
     setTheme((prev) => {
       const newTheme = prev === 'light' ? 'dark' : 'light';
       document.documentElement.setAttribute('data-theme', newTheme);
       localStorage.setItem('theme', newTheme);
       return newTheme;
     });
-  }, []);
+  };
 
   useEffect(() => {
     document.documentElement.setAttribute('data-theme', theme);
@@ -43,14 +43,4 @@ const ThemeContextProvider = ({ children }: IProps) => {
   };
 
   return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>;
-};
-
-export default ThemeContextProvider;
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (context === null) {
-    throw new Error('Context must be used within a context provider');
-  }
-  return context;
 };
