--- src/features/router/index.css
@@ -1,9 +1,0 @@
-.container {
-    padding: 20px;
-}
-
-@media (min-width: 768px) {
-    .container {
-        padding: 40px;
-    }
-}
