+++ src/features/router/index.tsx
@@ -9,19 +9,15 @@ import { MenuButton } from 'src/components/menuButton/menuButton';
 
 import { ABOUT_US_PATH, CONTACT_US_PATH, FALLBACK_404_ROUTE, HISTORY_OBSERVER_PATH } from './path';
 
-import './index.css';
-
 const Router = () => (
   <BrowserRouter>
-    <div className="container">
-      <MenuButton />
-      <Routes>
-        <Route path={HISTORY_OBSERVER_PATH + ':data?'} element={<HistoryPage />} />
-        <Route path={ABOUT_US_PATH} element={<AboutProjectPage />} />
-        <Route path={FALLBACK_404_ROUTE} element={<Page404 />} />
-        <Route path={CONTACT_US_PATH} element={<ContactUsPage />} />
-      </Routes>
-    </div>
+    <MenuButton />
+    <Routes>
+      <Route path={HISTORY_OBSERVER_PATH + ':data?'} element={<HistoryPage />} />
+      <Route path={ABOUT_US_PATH} element={<AboutProjectPage />} />
+      <Route path={FALLBACK_404_ROUTE} element={<Page404 />} />
+      <Route path={CONTACT_US_PATH} element={<ContactUsPage />} />
+    </Routes>
   </BrowserRouter>
 );
